pyproject.toml
[tool.poetry]
name = "tortally"
version = "0.1.0"
description = "...awesome!"
authors = ["jon <jonathan.n@u.nus.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.26.0"
pandas = "^2.1.0"
numpy = "^1.25.2"
transformers = "^4.33.0"
asyncio = "^3.4.3"
langchain = "^0.0.281"
openai = "^0.28.0"
python-dotenv = "^1.0.0"
pyngrok = "^6.0.0"
torch = "^2.0.1"
rich = "^13.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
